Fix an edge case with ssh-agent on Mac.
So it seems like MacOS also sets SSH_AUTH_SOCK even when just running
locally. So we can't use that test alone to see if we're on a forwarded
agent. But we aren't going to be doing the forward yubikey foo on mac. So
if we're on a Mac we can keychain up.